+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Fixes
 
+* Fixed error when OpenWeather doesn't provide a description for a weather condition.
 * Fixed index-out-of-bounds crash when using the temperature color scale with very high temperatures.
 
 ## [0.6.5] - 2021-11-26